refactor: App 컴포넌트의 라우팅 이벤트 로직 분리 및 스토어 내부의 함수에서 필요 없는 매개변수 삭제
